os/fs: move btrfs_ioctl.h to os/fs
