+++ src/common/sync_filesystem.h
@@ -20,7 +20,7 @@
 #if defined(__linux__)
 #include <sys/ioctl.h>
 #include <syscall.h>
-#include "../os/btrfs_ioctl.h"
+#include "os/fs/btrfs_ioctl.h"
 #endif
 
 inline int sync_filesystem(int fd)
